--- sheerwater_benchmarking/data/__init__.py
@@ -1,16 +1,0 @@
-"""Data functions for all parts of the data pipeline."""
-from .era5 import (single_era5, single_era5_cleaned, era5_rolled, era5_cds, era5_agg,
-                   salient_era5_raw, salient_era5)
-from .masks import (land_sea_mask)
-
-# Use __all__ to define what is part of the public API.
-__all__ = [
-    single_era5,
-    single_era5_cleaned,
-    era5_rolled,
-    era5_cds,
-    era5_agg,
-    salient_era5_raw,
-    salient_era5,
-    land_sea_mask,
-]
