+++ sheerwater_benchmarking/models/salient.py
@@ -1,6 +1,28 @@
 """Pulls Salient Predictions S2S forecasts from the Salient API."""
+# import numpy as np
+# import xarray as xr
+import salientsdk as sk
+
+from sheerwater_benchmarking.utils import cacheable, salient_auth, get_dates, get_grid
+from sheerwater_benchmarking.utils.general_utils import get_dates
+# from sheerwater_benchmarking.utils.model_utils import get_salient_loc
 
 
+@salient_auth
+def get_salient_loc(grid):
+    """Get and upload the location object for the Salient API."""
+    if grid != "africa0_25":
+        raise NotImplementedError("Only the African 0.25 grid is supported.")
+
+    # Upload location shapefile to Salient backend
+    lons, lats, _ = get_grid(grid)
+    coords = [(lons[0], lats[0]), (lons[-1], lats[0]), (lons[-1], lats[-1]), (lons[0], lats[-1])]
+    loc = sk.Location(shapefile=sk.upload_shapefile(
+        coords=coords,
+        geoname="all_africa",  # the full African continent
+        force=True))
+
+    return loc
 
 # @salient_auth
 # @cacheable(data_type='array',
