+++ sheerwater_benchmarking/reanalysis/__init__.py
@@ -1,0 +1,10 @@
+"""Data functions for all parts of the data pipeline."""
+from .era5 import era5, era5_rolled, era5_agg, salient_era5
+
+# Use __all__ to define what is part of the public API.
+__all__ = [
+    era5,
+    era5_rolled,
+    era5_agg,
+    salient_era5,
+]
