+++ sheerwater_benchmarking/utils/__init__.py
@@ -1,9 +1,9 @@
 """Utility functions for benchmarking."""
 from .caching import cacheable
-from .secrets import (cdsapi_secret, ecmwf_secret, salient_secret, salient_auth)
-from .data_utils import (apply_mask, roll_and_agg, regrid)
-from .general_utils import (get_grid, get_variable)
 from .remote import dask_remote
+from .secrets import cdsapi_secret, ecmwf_secret, salient_secret, salient_auth
+from .data_utils import apply_mask, roll_and_agg, regrid
+from .general_utils import get_grid, get_variable
 
 # Use __all__ to define what is part of the public API.
 __all__ = [
