+++ sheerwater_benchmarking/utils/general_utils.py
@@ -52,18 +52,16 @@ def dt_to_string(dt):
     return datetime.strftime(dt, DATETIME_FORMAT)
 
 
-valid_forecast_dates = {
-    "reforecast": {
-        "ecmwf": (string_to_dt("2015-05-14"), datetime.today(), "monday/thursday"),
-    },
-    "forecast": {
-        "ecmwf": (string_to_dt("2015-05-14"), datetime.today(), "monday/thursday"),
-    },
-}
-
-
 def is_valid_forecast_date(model, forecast_type, forecast_date):
     """Checks if the forecast date is valid for the given model and forecast type."""
+    valid_forecast_dates = {
+        "reforecast": {
+            "ecmwf": (string_to_dt("2015-05-14"), datetime.today(), "monday/thursday"),
+        },
+        "forecast": {
+            "ecmwf": (string_to_dt("2015-05-14"), datetime.today(), "monday/thursday"),
+        },
+    }
     assert isinstance(forecast_date, datetime)
     try:
         return forecast_date in generate_dates_in_between(
@@ -97,105 +95,6 @@ def generate_dates_in_between(first_date, last_date, date_frequency):
                                   (frequency_to_int[date_frequency])) + 1,)
         ]
         return dates
-
-
-def printf(str):
-    """Calls print on given argument and then flushes stdout buffer.
-
-    Ensures that printed message is displayed right away
-    """
-    print(str, flush=True)
-
-
-def print_fail(message="FAIL", verbose=True, skip_line_before=True,
-               skip_line_after=True, bold=False):
-    """Print message in purple."""
-    if verbose:
-        string_before = "\n" if skip_line_before else ""
-        string_after = "\n" if skip_line_after else ""
-        if bold:
-            print(
-                f"{string_before}\x1b[1;30;45m[ {message} ]\x1b[0m{string_after}")
-        else:
-            print(f"{string_before}\x1b[35m{message}\x1b[0m{string_after}")
-
-
-def print_error(message="ERROR", verbose=True, skip_line_before=True,
-                skip_line_after=True, bold=False):
-    """Print message in red."""
-    if verbose:
-        string_before = "\n" if skip_line_before else ""
-        string_after = "\n" if skip_line_after else ""
-        if bold:
-            print(
-                f"{string_before}\x1b[1;30;41m[ {message} ]\x1b[0m{string_after}")
-        else:
-            print(f"{string_before}\x1b[31m{message}\x1b[0m{string_after}")
-
-
-def print_warning(message="WARNING", verbose=True, skip_line_before=True,
-                  skip_line_after=True, bold=False):
-    """Print message in yellow."""
-    if verbose:
-        string_before = "\n" if skip_line_before else ""
-        string_after = "\n" if skip_line_after else ""
-        if bold:
-            print(
-                f"{string_before}\x1b[1;30;43m[ {message} ]\x1b[0m{string_after}")
-        else:
-            print(f"{string_before}\x1b[33m{message}\x1b[0m{string_after}")
-
-
-def print_ok(message="OK", verbose=True, skip_line_before=True, skip_line_after=True, bold=False):
-    """Print message in green."""
-    if verbose:
-        string_before = "\n" if skip_line_before else ""
-        string_after = "\n" if skip_line_after else ""
-        if bold:
-            print(
-                f"{string_before}\x1b[1;30;42m[ {message} ]\x1b[0m{string_after}")
-        else:
-            print(f"{string_before}\x1b[32m{message}\x1b[0m{string_after}")
-
-
-def print_info(message, verbose=True):
-    """Print message if verbose."""
-    if verbose:
-        print(message)
-
-
-class TLSAdapter(requests.adapters.HTTPAdapter):
-    """Transport adapter that allows us to use TLSv1.2."""
-
-    def init_poolmanager(self, connections, maxsize, block=False):
-        """Create and initialize the urllib3 PoolManager."""
-        ctx = ssl.create_default_context()
-        ctx.set_ciphers('DEFAULT@SECLEVEL=1')
-        self.poolmanager = poolmanager.PoolManager(
-            num_pools=connections,
-            maxsize=maxsize,
-            block=block,
-            ssl_version=ssl.PROTOCOL_TLS,
-            ssl_context=ctx)
-
-
-def download_url(url, timeout=600, retry=3, cookies={}):
-    """Download URL, waiting some time between retries."""
-    r = None
-    session = requests.session()
-    session.mount('https://', TLSAdapter())
-
-    for i in range(retry):
-        try:
-            r = session.get(url, timeout=timeout, cookies=cookies)
-            return r
-        except requests.exceptions.Timeout as e:
-            # Wait until making another request
-            if i == retry - 1:
-                raise e
-            print(f"Request to url {url} has timed out. Trying again...")
-            time.sleep(3)
-    print(f"Failed to retrieve file after {retry} attempts. Stopping...")
 
 
 def get_dates(start_time, end_time, stride="day", return_string=False):
@@ -283,7 +182,19 @@ def get_grid(region_id):
         longitudes = np.arange(0.125, 360, 0.25)
         latitudes = np.arange(-89.875, 90, 0.25)
         grid_size = 0.25
+    elif region_id == "africa0_25":
+        longitudes = np.arange(0.125, 360, 0.25)
+        latitudes = np.arange(-89.875, 90, 0.25)
+        grid_size = 0.25
+    elif region_id == "africa1_5":
+        longitudes = np.arange(-26.0, 73.0, 1.5)
+        latitudes = np.arange(-35.0, 38.0, 1.5)
+        grid_size = 1.5
+    elif region_id == "africa0_25":
+        longitudes = np.arange(-26.0, 73.0, 0.25)
+        latitudes = np.arange(-35.0, 38.0, 0.25)
+        grid_size = 0.25
     else:
         raise NotImplementedError(
-            "Only grids global1_5, us1_0 and us1_5 have been implemented.")
+            f"Grid {region_id} has not been implemented.")
     return longitudes, latitudes, grid_size
